requirements.txt
aiogram==3.3.0
aiohttp>=3.8.0
python-dotenv
EOFcat > .env.example << 'EOF'
TELEGRAM_BOT_TOKEN=YOUR_TELEGRAM_TOKEN_HERE
GIGACHAT_CLIENT_ID=YOUR_CLIENT_ID_HERE
GIGACHAT_CLIENT_SECRET=YOUR_CLIENT_SECRET_HERE
GIGACHAT_SCOPE=GIGACHAT_API_PERS
